# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /income:
    get:
      operationId: listIncome
      tags: [income]
      summary: List all income entries
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: frequency
          schema:
            type: string
            enum: [monthly, annually, one_time]
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [active, cancelled]
          required: false
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeEntry"
    post:
      operationId: createIncome
      tags: [income]
      summary: Create a new income entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeBody"
      responses:
        "201":
          description: Created income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /income/summary:
    get:
      operationId: getIncomeSummary
      tags: [income]
      summary: Get income totals grouped by frequency and category
      parameters:
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Income summary totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSummary"
  /income/{id}:
    get:
      operationId: getIncomeEntry
      tags: [income]
      summary: Get a single income entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIncomeEntry
      tags: [income]
      summary: Update an income entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncomeBody"
      responses:
        "200":
          description: Updated income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteIncomeEntry
      tags: [income]
      summary: Delete an income entry permanently
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /income/{id}/cancel:
    patch:
      operationId: cancelIncomeEntry
      tags: [income]
      summary: Cancel an income entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CategorySummary:
      type: object
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
      required:
        - category
        - total
        - count
...
    IncomeEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        status:
          type: string
          enum: [active, cancelled]
        notes:
          type: ["string", "null"]
        receivedDate:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - amount
        - currency
        - category
        - frequency
        - status
        - notes
        - receivedDate
        - renewalDate
        - createdAt
        - updatedAt
    CreateIncomeBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        notes:
          type: ["string", "null"]
        receivedDate:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
      required:
        - name
        - amount
        - category
        - frequency
    UpdateIncomeBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        status:
          type: string
          enum: [active, cancelled]
        notes:
          type: ["string", "null"]
        receivedDate:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
    IncomeSummary:
      type: object
      properties:
        totalMonthly:
          type: number
        totalAnnually:
          type: number
        totalOneTime:
          type: number
        totalAllAnnualized:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
        activeCount:
          type: integer
        cancelledCount:
          type: integer
      required:
        - totalMonthly
        - totalAnnually
        - totalOneTime
        - totalAllAnnualized
        - byCategory
        - activeCount
        - cancelledCount
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - createdAt
        - updatedAt
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - updatedAt
        - messages
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content