# lib/api-zod/src/generated/types/createIncomeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateIncomeBodyFrequency } from "./createIncomeBodyFrequency";

export interface CreateIncomeBody {
  name: string;
  amount: number;
  currency?: string;
  category: string;
  frequency: CreateIncomeBodyFrequency;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  receivedDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
}