# lib/api-zod/src/generated/types/incomeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IncomeEntryFrequency } from "./incomeEntryFrequency";
import type { IncomeEntryStatus } from "./incomeEntryStatus";

export interface IncomeEntry {
  id: number;
  name: string;
  amount: number;
  currency: string;
  category: string;
  frequency: IncomeEntryFrequency;
  status: IncomeEntryStatus;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  receivedDate: string | null;
  /** @nullable */
  renewalDate: string | null;
  createdAt: string;
  updatedAt: string;
}