# lib/api-zod/src/generated/types/incomeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySummary } from "./categorySummary";

export interface IncomeSummary {
  totalMonthly: number;
  totalAnnually: number;
  totalOneTime: number;
  totalAllAnnualized: number;
  byCategory: CategorySummary[];
  activeCount: number;
  cancelledCount: number;
}