# lib/api-zod/src/generated/types/updateIncomeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateIncomeBodyFrequency } from "./updateIncomeBodyFrequency";
import type { UpdateIncomeBodyStatus } from "./updateIncomeBodyStatus";

export interface UpdateIncomeBody {
  name?: string;
  amount?: number;
  currency?: string;
  category?: string;
  frequency?: UpdateIncomeBodyFrequency;
  status?: UpdateIncomeBodyStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  receivedDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
}